123140192_Abi Sholihan_Praktikum 7_FastAPI_NEW/requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
python-jose
passlib[bcrypt]
bcrypt==3.2.2
python-multipart